Struct underscore_vm::Chunk

pub struct Chunk {
    pub constants: Vec<u8>,
    // some fields omitted
}

Fields
constants: Vec<u8>
Methods

impl Chunk

pub fn new() -> Self

pub fn write<T: Into<u8>>(&mut self, byte: T, line: usize)

Write a single byte to this chunk

pub fn write_values(&mut self, bytes: &[u8], line: usize)

Write multiple bytes to this chunk

pub fn add_constant(&mut self, bytes: &[u8], line: usize) -> usize

Adds a constant to the constant pool. Constant must be in bytes. Returns the offset at which the constant is stored

pub fn dissassemble(&mut self, name: &str)

pub fn dissassemble_instruction(&self, offset: usize) -> usize

pub fn constant_instruction(
&self,
name: &str,
size: usize,
offset: usize
) -> usize

Pulls out the OpCode::Constant and prints that out Matches on the instruction and uses that pointer offset for were the value is stored